+++ package.json
@@ -64,6 +64,7 @@
     "Jason Sturges <jason@jasonsturges.com> (https://github.com/jasonsturges)",
     "Jean-Elie Barjonet (https://github.com/jebarjonet)",
     "Jeff Francisco (https://github.com/j-francisco)",
+    "Jesse Mandel (https://github.com/supergibbs)",
     "Jim Liu <junminliu@gmail.com> (https://github.com/JimLiu)",
     "Joe (https://github.com/joemcadams)",
     "Joe Woynillowicz (https://github.com/CoderJoeIO)",
