+++ src/NavDropdown.js
@@ -1,39 +1,8 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import { mapToCssModules } from './utils';
+import { warnOnce } from './utils';
 import Dropdown from './Dropdown';
 
-const propTypes = {
-  children: PropTypes.node,
-  tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
-  className: PropTypes.string,
-  cssModule: PropTypes.object,
-};
-
-const defaultProps = {
-  tag: 'li'
-};
-
-const NavDropdown = (props) => {
-  const {
-    className,
-    cssModule,
-    tag: Tag,
-    ...attributes
-  } = props;
-
-  const classes = mapToCssModules(classNames(
-    className,
-    'nav-item'
-  ), cssModule);
-
-  return (
-    <Dropdown {...attributes} tag={Tag} className={classes} />
-  );
-};
-
-NavDropdown.propTypes = propTypes;
-NavDropdown.defaultProps = defaultProps;
-
-export default NavDropdown;
+export default function NavDropdown(props) {
+  warnOnce('The "NavDropdown" component has been deprecated.\nPlease use component "Dropdown" with nav prop.');
+  return <Dropdown nav {...props} />;
+}
