+++ src/__tests__/Uncontrolled.spec.js
@@ -4,12 +4,10 @@ import {
   Alert,
   ButtonDropdown,
   Dropdown,
-  NavDropdown,
   Tooltip,
   UncontrolledAlert,
   UncontrolledButtonDropdown,
   UncontrolledDropdown,
-  UncontrolledNavDropdown,
   UncontrolledTooltip,
 } from '../';
 
@@ -82,30 +80,6 @@ describe('UncontrolledDropdown', () => {
     const instance = dropdown.instance();
     instance.toggle();
     expect(dropdown.prop('isOpen')).toBe(true);
-  });
-});
-
-describe('UncontrolledNavDropdown', () => {
-  it('should be an NavDropdown', () => {
-    const navDropdown = shallow(<UncontrolledNavDropdown>Yo!</UncontrolledNavDropdown>);
-    expect(navDropdown.type()).toBe(NavDropdown);
-  });
-
-  it('should have isOpen default to false', () => {
-    const navDropdown = shallow(<UncontrolledNavDropdown>Yo!</UncontrolledNavDropdown>);
-    expect(navDropdown.prop('isOpen')).toBe(false);
-  });
-
-  it('should have toggle function', () => {
-    const navDropdown = shallow(<UncontrolledNavDropdown>Yo!</UncontrolledNavDropdown>);
-    expect(navDropdown.prop('toggle')).toEqual(expect.any(Function));
-  });
-
-  it('should toggle isOpen when toggle is called', () => {
-    const navDropdown = shallow(<UncontrolledNavDropdown>Yo!</UncontrolledNavDropdown>);
-    const instance = navDropdown.instance();
-    instance.toggle();
-    expect(navDropdown.prop('isOpen')).toBe(true);
   });
 });
